Update ReservationTime to have start and end
We need to also track when the reservation ends. I believe this is
better than storing the start and the duration of a reservation, because
now we can query the database for ReservationTimes that have a start or
an end in a timespan, and the end has already been calculated at the
creation.